third_party/blink/web_tests/flag-specific/enable-skia-graphite/external/wpt/webrtc-encoded-transform/script-transform-sendKeyFrameRequest.https-expected.txt

This is a testharness.js-based test.
[FAIL] sendKeyFrameRequest does not reject when the receiver is negotiated inactive, and resumes succeeding when negotiated back to active
  assert_equals: expected "success" but got "failure: InvalidStateError"
Harness: the test ran to completion.
